+++ core/i18n/format.js
