+++ core/i18n/index.ts
@@ -5,5 +5,5 @@
  */
 
 export * from './date/index.js'
-export * from './date/format.js'
+export * from './format.js'
 export * from './duration.js'
